refactor(frontend): move calcul indemnite steps into a folder
